Reduce stack_len to 32-bit

+++ examples/cpp/pyperf/PyPerfBPFProgram.cc
@@ -156,7 +156,7 @@ struct event {
   int32_t kernel_stack_id;
   // instead of storing symbol name here directly, we add it to another
   // hashmap with Symbols and only store the ids here
-  uint64_t stack_len;
+  int32_t stack_len;
   int32_t stack[STACK_MAX_LEN];
 };
